moved Dataset and GeosyncDBEntities to separate files
